+++ src/test/punctuation/single-quotes.test.js
@@ -20,7 +20,7 @@ describe('Single quotes in default language (en)\n', () => {
 
 
 	Object.keys(testCase).forEach((key) => {
-		it("should replace multiple dots with an ellipsis", () => {
+		it("should fix single quotes, primes and apostrophes in English", () => {
 			assert.equal(fixSingleQuotesPrimesAndApostrophes(key, "en"), testCase[key]);
 		});
 	});
@@ -39,7 +39,7 @@ describe('Single quotes in Slovak and Czech language (sk, cs)\n', () => {
 
 
 	Object.keys(testCase).forEach((key) => {
-		it("should replace multiple dots with an ellipsis", () => {
+		it("should fix single quotes, primes and apostrophes in Slovak/Czech", () => {
 			assert.equal(fixSingleQuotesPrimesAndApostrophes(key, "sk"), testCase[key]);
 			assert.equal(fixSingleQuotesPrimesAndApostrophes(key, "cs"), testCase[key]);
 		});
@@ -60,7 +60,7 @@ describe('Single quotes in Rusyn language (rue)\n', () => {
 
 
 	Object.keys(testCase).forEach((key) => {
-		it("should replace multiple dots with an ellipsis", () => {
+		it("should fix single quotes, primes and apostrophes in Rusyn", () => {
 			assert.equal(fixSingleQuotesPrimesAndApostrophes(key, "rue"), testCase[key]);
 		});
 	});